Clean up sc.datasets tab completion
Moved definition in sc.datasets to a private module so the tab completion namespace doesn't get polluted by irrelevant names
